thru CH5 L6: Create Chirp
Add a POST /api/chirps handler. Delete the /api/validate_chirp endpoint that we created before, but port all that logic into this one. Users should not be allowed to create invalid chirps! If the chirp is valid, you should save it in the database. You'll need a new up/down migration for this table. If creating the record goes well, respond with a 201 status code and the full chirp resource.