Authorization (#2)
* created authenticateUser helper function for request handlers to use

* added a UpdateEmailAndPasswordByUserId query

* Add a PUT /api/users endpoint so that users can update their own (but not others') email and password.

Add a PUT /api/users endpoint so that users can update their own (but not others') email and password. It requires:
An access token in the header
A new password and email in the request body
Hash the password, then update the hashed password and the email for the authenticated user in the database. Respond with a 200 if everything is successful and the newly updated User resource (omitting the password of course).
If the access token is malformed or missing, respond with a 401 status code.

* DeleteChirpById query added

* Add a new DELETE /api/chirps/{chirpID}

Add a new DELETE /api/chirps/{chirpID} route to your server that deletes a chirp from the database by its id.
This is an authenticated endpoint, so be sure to check the token in the header. Only allow the deletion of a chirp if the user is the author of the chirp.
If they are not, return a 403 status code.
If the chirp is deleted successfully, return a 204 status code.
If the chirp is not found, return a 404 status code.